refactor private tree class out - avoid circular imports
